Components: Handle new  connectedGarden provider
