+++ src/components/MainView.js
@@ -7,19 +7,20 @@ import Header from './Header/Header'
 import Layout from './Layout'
 import GlobalPreferences from './Garden/Preferences/GlobalPreferences'
 import Sidebar from './Sidebar/Sidebar'
+import { useConnectedGarden } from '@providers/ConnectedGarden'
+import { useGardenState } from '@providers/GardenState'
 import usePreferences from '@hooks/usePreferences'
-import { useGardens } from '@/providers/Gardens'
-import { useGardenState } from '@/providers/GardenState'
 
 function MainView({ children }) {
   const { pathname } = useLocation()
   const { below } = useViewport()
-  const { connectedGarden } = useGardens()
+  const connectedGarden = useConnectedGarden()
 
   const [openPreferences, closePreferences, preferenceOption] = usePreferences()
 
   let loadingGardenState = true
   if (connectedGarden) {
+    // TODO: Refactor
     // eslint-disable-next-line react-hooks/rules-of-hooks
     const { loading } = useGardenState()
     loadingGardenState = loading
