+++ src/hooks/useGardenTokenIcon.js
@@ -1,7 +1,7 @@
-import { useGardens } from '@providers/Gardens'
+import { useConnectedGarden } from '@providers/ConnectedGarden'
 import { getGardenTokenIcon } from '../utils/token-utils'
 
 export default function useGardenTokenIcon(token) {
-  const { connectedGarden } = useGardens()
+  const connectedGarden = useConnectedGarden()
   return getGardenTokenIcon(connectedGarden, token)
 }
